refactor: reorganize resolver structure and introduce new read/write handling
