+++ api/src/cmd/compiler/index.ts
@@ -3,5 +3,5 @@ export * from "./errors.js";
 export * from "./unwrap.js";
 export * from "./identify.js";
 export * from "./parsers/index.js";
-export * from "./resolver.js";
+export * from "./resolver/resolver.js";
 export * from "./analysis.js";
