+++ api/src/cmd/compiler/resolver/index.ts
@@ -1,0 +1,1 @@
+export { CompilerCommandResolver } from "./resolver.js";
